+++ src/components/carousel/index.ts
@@ -1,9 +1,6 @@
 import Carousel from 'react-slick';
-
-export { default as CarouselDots } from './carousel-dots';
-export { default as CarouselArrows } from './carousel-arrows';
-export { default as CarouselArrowIndex } from './carousel-arrow-index';
-
-export { default as useCarousel } from './use-carousel';
-
+export { default as CarouselDots } from './components/CarouselDots.tsx';
+export { default as CarouselArrows } from './components/CarouselArrows.tsx';
+export { default as CarouselArrowIndex } from './components/CarouselArrowIndex.tsx';
+export { default as useCarousel } from '@src/hooks/components/use-carousel';
 export default Carousel;