+++ src/sections/finance/components/finance-earn-tokens.tsx
@@ -15,7 +15,10 @@ import { alpha, useTheme } from '@mui/material/styles';
 import Earn from '@src/assets/illustrations/earn.svg';
 import Image from '@src/components/image';
 
-interface FinanceEarnTokensProps extends CardProps {
+// Global constants
+import { GLOBAL_CONSTANTS } from '@src/config-global';
+
+export interface FinanceEarnTokensProps extends CardProps {
   lgUp: boolean;
 }
 
@@ -23,8 +26,7 @@ const FinanceEarnTokens: FC<FinanceEarnTokensProps> = ({ sx, lgUp, ...other }) =
   const theme = useTheme();
 
   const handleClick = () => {
-    // TODO move to envs
-    window.open('https://zealy.io/cw/watchit', '_BLANK');
+    window.open(GLOBAL_CONSTANTS.EARN_TOKEN_SERVICE_URL, '_BLANK');
   };
 
   return (
@@ -127,6 +129,7 @@ const FinanceEarnTokens: FC<FinanceEarnTokensProps> = ({ sx, lgUp, ...other }) =
             </Button>
           </Stack>
           <Stack
+            data-testid="earn-tokens-image"
             flexGrow={1}
             justifyContent="center"
             sx={{